wip: extract logic into leptos_macro_core
